1.添加种子数据初始化模块 SeedDataInitModule，用于初始化系统运行的一些必需数据。2.对系统数据的更新与删除进行限制。3.修复一些CreateScope之后对scope.ServiceProvider使用不正确的问题。

+++ src/OSharp.Permissions/Identity/RoleStoreBase.cs
@@ -102,6 +102,10 @@ namespace OSharp.Identity
             ThrowIfDisposed();
             Check.NotNull(role, nameof(role));
 
+            if (role.IsSystem)
+            {
+                return new IdentityResult().Failed($"角色“{role.Name}”是系统角色，不能更新");
+            }
             await _roleRepository.UpdateAsync(role);
             return IdentityResult.Success;
         }
@@ -118,6 +122,10 @@ namespace OSharp.Identity
             ThrowIfDisposed();
             Check.NotNull(role, nameof(role));
 
+            if (role.IsSystem)
+            {
+                return new IdentityResult().Failed($"角色“{role.Name}”是系统角色，不能删除");
+            }
             await _roleRepository.DeleteAsync(role);
             return IdentityResult.Success;
         }
